backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cvzone>=1.6.1",
    "fastapi[standard]>=0.115.12",
    "mediapipe>=0.10.14",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pydantic-settings>=2.8.1",
]

[dependency-groups]
dev = [
    "dotenv>=0.9.9",
]
